add cnn test

+++ cifar10/train.py
@@ -3,13 +3,15 @@ import torch.nn as nn
 import torchvision
 import torchvision.transforms as transforms
 import torch.optim as optim
-import time, csv
+import time, csv, math
 
-from torch.optim.lr_scheduler import StepLR
+from torch.optim.lr_scheduler import StepLR, LambdaLR
 from pathlib import Path
 
 from utils import *
-from resnet18k import make_resnet18k
+from model_resnet18k import make_resnet18k
+from model_mcnn import CNN
+
 
 def get_hparams():
     return [
@@ -19,6 +21,8 @@ def get_hparams():
         ['-L', '--ldecay', int,   1000],
         ['-n', '--width',  int,   10],
         ['-s', '--noise',  int,   1500],
+        ['-o', '--optim',  str,   "adam"],
+        ['-m', '--model',  str,   "resnet18"],
     ]
 
 
@@ -27,7 +31,8 @@ def get_arguments():
         ['-t', '--trials', int,   5],
         ['-e', '--epochs', int,   1000],
         [None, '--logdir', Path,  Path('out')],
-        [None, '--gpu',    int,   -1]
+        [None, '--gpu',    int,   -1],
+        [None, '--saveall'],
     ]
 
 
@@ -36,7 +41,11 @@ def build_parser():
     parser = argparse.ArgumentParser()
     hparams, arguments = get_hparams(), get_arguments()
     for line in hparams+arguments:
-        parser.add_argument(*filter(None, line[:2]), type=line[-2], default=line[-1])
+        if len(line) == 2:
+            parser.add_argument(*filter(None, line[:2]), action='store_true')
+        else:
+            parser.add_argument(*filter(None, line[:2]), type=line[-2], default=line[-1])
+
     return parser
 
 
@@ -66,15 +75,12 @@ def train(model, trainloader, optimizer, criterion, flags):
     model.train()
     train_loss, correct, total = 0, 0, 0
     for batch_idx, (inputs, targets) in enumerate(trainloader):
-        targets_onehot = torch.FloatTensor(targets.size(0), 10)
         if flags.gpu >= 0:
             inputs, targets = inputs.cuda(), targets.cuda()
             targets = targets.cuda()
-        targets_onehot.zero_()
-        targets_onehot.scatter_(1, targets.view(-1, 1).long(), 1)
         optimizer.zero_grad()
         outputs = model(inputs)
-        loss = criterion(outputs, targets_onehot)
+        loss = criterion(outputs, targets)
         loss.backward()
         optimizer.step()
         train_loss += loss.item() * outputs.numel()
@@ -89,14 +95,11 @@ def valid(model, validloader, optimizer, criterion, flags):
     test_loss, correct, total = 0, 0, 0
     with torch.no_grad():
         for batch_idx, (inputs, targets) in enumerate(validloader):
-            targets_onehot = torch.FloatTensor(targets.size(0), 10)
             if flags.gpu >= 0:
                 inputs, targets = inputs.cuda(), targets.cuda()
                 targets = targets.cuda()
-            targets_onehot.zero_()
-            targets_onehot.scatter_(1, targets.view(-1, 1).long(), 1)
             outputs = model(inputs)
-            loss = criterion(outputs, targets_onehot)
+            loss = criterion(outputs, targets)
             test_loss += loss.item() * outputs.numel()
             _, predicted = outputs.max(1)
             correct += predicted.eq(targets).sum().item()
@@ -124,15 +127,25 @@ def main(flags):
         trainloader = torch.utils.data.DataLoader(trainsubset, batch_size=128, shuffle=True)
 
         # model & loss
-        criterion = nn.MSELoss(reduction='mean')
-        model = make_resnet18k(flags.width, num_classes=10)
+        criterion = nn.CrossEntropyLoss()
+        if flags.model == 'resnet18':
+            model = make_resnet18k(flags.width, num_classes=10)
+        elif flags.model == 'cnn':
+            model = CNN(flags.width, num_classes=10)
+
         # move to GPU
-        if args.gpu >= 0:
+        if flags.gpu >= 0:
             criterion = criterion.cuda()
             model = model.cuda()
         
-        optimizer = optim.Adam(model.parameters(), lr=flags.lrate, weight_decay=flags.wdecay)
-        scheduler = StepLR(optimizer, step_size=flags.ldecay, gamma=0.1)
+        if flags.optim == 'adam':
+            optimizer = optim.Adam(model.parameters(), lr=flags.lrate, weight_decay=flags.wdecay)
+            scheduler = StepLR(optimizer, step_size=flags.ldecay, gamma=0.1)
+        elif flags.optim == 'sgd':
+            optimizer = optim.SGD(model.parameters(), lr=flags.lrate, momentum=0, weight_decay=flags.wdecay)
+            trainset_size = 50_000
+            lr_lambda = lambda epoch: flags.lrate / math.sqrt(1 + (epoch*trainset_size)//512)
+            scheduler = LambdaLR(optimizer, lr_lambda=lr_lambda, last_epoch=-1)
 
         with open(logdir / f'trainlog_{trial}_{timestamp}.csv', 'w') as csvf:
             best_acc, best_epoch = 0, -1
@@ -149,6 +162,10 @@ def main(flags):
                 })
                 csvf.flush()
 
+                if flags.saveall:
+                    torch.save(model.state_dict(), logdir / f'model_{token}_{trial}_{epoch}.pkl')
+                    continue
+                
                 if epoch >= 10 and valid_acc > best_acc:
                     prev = logdir / f'model_{token}_{trial}_{best_epoch}.pkl'
                     if prev.exists(): prev.unlink()
